Changing the loss function

+++ trainer/trainer.py
@@ -215,7 +215,7 @@ class Trainer(GenericTrainer):
                     y_onehot[old_classes_indices] = outpu2.data
 
             # loss = F.binary_cross_entropy(output, Variable(y_onehot))
-            loss = F.multilabel_margin_loss(output, Variable(y_onehot))
+            loss = F.multilabel_margin_loss(output, Variable(y_onehot.long()))
             loss.backward()
             self.optimizer.step()
             bar.update(int(float(batch_idx+1)/float(len(self.train_data_iterator))*100))
